+++ clx_invoice_reports/models/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo, CLx Media
+# See LICENSE file for full copyright & licensing details.
+from . import res_company
+from . import account_move