+++ clx_task_management/models/__init__.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo, CLx Media
+# See LICENSE file for full copyright & licensing details.
+from . import main_task
+from . import sub_task
+from . import team
+from . import request_form
+from . import project
+from . import sale_order_line
+from . import sale_order
+from . import res_config_settings